requirements.txt
# Backend dependencies
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
aiohttp==3.8.6
pymupdf==1.23.3
jsonschema==4.19.1
python-dotenv==1.0.0

# Frontend dependencies
streamlit==1.27.2
requests==2.31.0

# Note: sqlite3 is part of Python's standard library, no need to install it separately
